requirements_top_level.txt
boto3
pyfakefs
pyhcl
pylint
yara-python==3.6.3
